Move logserver to separate file
